pyproject.toml
[project]
name = "rustyrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.2",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
    "weaviate-client>=4.16.10",
]
